+++ app/(nav)/todos/[todoId]/note/[noteId]/page.tsx
@@ -1,5 +1,5 @@
 import { cookies } from 'next/headers';
-import NoteForm from '../../_view/NoteForm';
+import NoteFormSections from '../../_view/NoteFormSections';
 import { Note } from '@/lib/types/todo';
 
 export default async function Page({ params }: { params: { noteId: string } }) {
@@ -15,7 +15,7 @@ export default async function Page({ params }: { params: { noteId: string } }) {
 
   return (
     <main className='lg:flex h-screen w-screen'>
-      <NoteForm title={title} content={content} linkUrl={linkUrl ?? ''} method='PATCH' noteId={noteId} />
+      <NoteFormSections title={title} content={content} linkUrl={linkUrl ?? ''} method='PATCH' noteId={noteId} />
     </main>
   );
 }
